replicate ability map

+++ include/sol/sol.h
@@ -54,7 +54,7 @@
 
 #define BITC(x)(1u << (x))
 #define UISCALE(x)                                                                                                     \
-    (x * min((float)Sol_GetState()->windowWidth / WINDOW_WIDTH, (float)Sol_GetState()->windowHeight / WINDOW_HEIGHT))
+    (x * min((double)Sol_GetState()->windowWidth / WINDOW_WIDTH, (double)Sol_GetState()->windowHeight / WINDOW_HEIGHT))
 
 SOLAPI int       Sol_Init(void *hwnd, void *hInstance);
 SOLAPI void      Sol_Tick(double dt, double time);
